ws-akka: move hardcoded config to config file
Note: this has some semantic changes, to follow the best practices
of using Typesafe Config in a library [1].

`ConfigFactory.load()` loads all `reference.conf` files, especially the
one that Akka provides themself. Also, environment and system variables
are laoded into the config.

As this is a library that may be used in a product with other akka
components, use a own config path "loci.communicator.ws.akka". This
allows to specify the akka parameters just for usage within this module,
and still using any globally defined akka configurations via `withFallback`.

[1] https://github.com/lightbend/config#using-the-library