CI: add workflow_dispatch trigger
This allows maintainers to trigger a workflow run from GitHub Actions,
optionally for a given commit or branch.

This can be used for example, when the CI for a longer PR fails, to
manually bisect which commit introduced the error.

The behavior for push and pull_request is unchanged (just a syntax change).