sbt: set crossScalaVersions to Nil on aggregating projects
There is a known limitation of how sbt-crossproject interacts with cross building projects
that have different values for `crossScalaVersions` [1].
Avoid running into future problems with the documented setting `crossScalaVersions` with Nil
on root project(s), which is the currently recommended value on aggregating projects [2].

Note: actually having different values for `crossScalaVersions` is not possible
with sbt & sbt-scalajs-crossproject.
Different values for `crossScalaVersions` are supposed to be possible with
sbt-projectmatrix, which is being considered to be in-sourced for sbt 2.0.
This change is already preparing to not run into issue when changing to
sbt-projectmatrix / upgrading to sbt 2.0 ["idea 8-A" in 3].

[1] https://www.github.com/sbt/sbt/issues/6193#issuecomment-738946050
[2] https://www.scala-sbt.org/1.x/docs/Cross-Build.html#Cross+building+a+project+statefully
[3] https://eed3si9n.com/sbt-2.0-ideas