CI: depend only on major version for coursier/setup-action
Not needed for anything, just to stay up to date.
Currently coursier/setup-action is at version 1.3.5.

GitHub Actions are typically just depending on a major version,
to avoid maintaining too many smaller version updates.

The dependency to a specific version was added for a Milestone release,
which made sense at the time, as perhaps some experimental support
was needed for Scala 3.